abi: the Hard geometry is double, a fractional block included
2.x took a fractional host buffer size (HostConfig::m_buffer_size is a
float; a plugin prepared a control-rate decoder with
samplesPerBlock / 2048.f), and the core has always handled one: the
LatencyCalculator rationalizes the block (1e-6 relative, denominator at
most 2^20) and sizes the latency, the slots and the rings for a host
that moves a sample once a whole one has accumulated. 3.x refused it
only at the boundary, where the geometry was uint32_t. It is double now
(Fares' decision, 2026-09-26: doubles, not a num/den pair), for the
model with no stream at the host rate, whose anchor block is a fraction
of a sample.

- The registry: block_min and block_max are double in
  anira_contract_create_hard and anira_contract_hard_set_geometry, and
  the read-back anira_contract_hard_geometry writes double* outs. No
  record carries the geometry, so the layout gate is untouched. The
  entry docs state the rules and the accounting.
- The rules, one function (geometry_refusal, handles.h / config.cpp)
  shared by both setters, the contract-file loader and the validator:
  both blocks 0 (the unset geometry of create and of the legacy
  contract) or both finite and > 0, block_min <= block_max, a rate of 0
  or more. A refusal is INVALID_ARGUMENT naming the value ("block_min
  nan must be finite and > 0 (block_max 1; both 0 leave the geometry
  unset)", "block_min 0.5 exceeds block_max 0.25"); set_geometry, which
  has no anira_error, gives the message to the boundary trace and
  leaves the geometry as it was. block_min 0 with a non-zero block_max,
  legal before, is refused.
- HardContract stores doubles; make_host_config hands block_max to the
  2.x HostConfig as the nearest float (a float until PR 10b makes the
  core native), from which rationalize recovers the rational the host
  meant.
- The flexible-window pin takes the rational of block_max and still
  requires a whole hop; its refusal names the rational ("gives a
  fractional hop for block_max 1/4").
- The contract file's block_min / block_max take any number. There is
  no contract writer to round-trip through (anira_contract_to_json does
  not exist), so the reader is the whole JSON side.
- anira.hpp: Hard::block_min / block_max and
  ContractHandle::hard_geometry are double, hard() reads doubles back;
  integer literals and uint32_t assignments keep compiling.
- The ANIRA_WAIT_CONTRACT budget of the pop twins reads the double
  directly.

The compat shim still refuses a fractional HostConfig here; the next
commits add the per-call documentation and the shim.

Verified: gen.py --check (62 files); gen.py --diff-against
v3.0.0-alpha.1 adds exactly two BREAKING lines,
anira_contract_create_hard and anira_contract_hard_set_geometry
"signature changed" (both since 0.1; anira_contract_hard_geometry is
since 0.2, no line), and abi/symbols-0.txt is unchanged at 227 promised
names. D (desktop-tests-debug) 990/990, the 986 of ff0c505 plus
AbiContract.TheGeometryRules, AbiJsonContract.FractionalGeometry,
AbiTranslate.AFractionalGeometryIsTheCalculatorsRational (1/3 and 0.25
on hops 1 and 4, and 2/3 on hop 4, through the float HostConfig give
exactly 1/3, 1/4, 1/12, 1/16 and 1/6 hops) and
AbiTranslate.AFractionalBlockMaxGivesAFractionalFlexibleHop.
clang-format 20.1.8 over the changed files.

+++ include/anira/anira.hpp
@@ -884,8 +884,11 @@ private:
 
 /// The real-time stream: host geometry, budget, warmup, miss policy, wait ratio.
 struct Hard {
-    uint32_t block_min = 0;
-    uint32_t block_max = 0;
+    /// The block range in elements of the anchor tensor: both 0 (unset) or both finite and
+    /// > 0, a fractional block included (0.25: one element every fourth call; the rules of
+    /// anira_contract_create_hard). An integer converts: `.block_max = 512`.
+    double block_min = 0;
+    double block_max = 0;
     double rate = 0;
     anira_budget_kind budget = ANIRA_BUDGET_MEASURED;
     std::chrono::nanoseconds budget_value{};  ///< Explicit only
@@ -1065,8 +1068,10 @@ public:
 
     // -- the setters, for a contract that exists only as a handle (loaded, legacy) --
 
-    /// Patches a Hard contract's stream geometry, e.g. of one loaded from a file.
-    ContractHandle& hard_geometry(uint32_t block_min, uint32_t block_max, double rate) {
+    /// Patches a Hard contract's stream geometry, e.g. of one loaded from a file; a block may
+    /// be fractional (anira_contract_create_hard). @throws Error{ANIRA_ERROR_INVALID_ARGUMENT}
+    /// for a geometry the rules refuse, Error{ANIRA_ERROR_WRONG_CONTRACT} on an Async handle.
+    ContractHandle& hard_geometry(double block_min, double block_max, double rate) {
         detail::check(anira_contract_hard_set_geometry(m_contract, block_min, block_max, rate),
                       "anira_contract_hard_set_geometry");
         return *this;
